Refactor: promote shared player types out of mpv package
Define LocalPlayer interface and shared AudioDevice/MediaInfo types in
backend/player so both mpv and a future alternative backend can implement
the same contract without importing each other.

- backend/player/equalizer.go: Equalizer interface + ISO10/15BandEqualizer
  moved from backend/player/mpv; mpv/equalizer.go is now type aliases
- backend/player/localplayer.go: LocalPlayer interface, AudioDevice, MediaInfo
- backend/player/mpv/player.go: implements player.LocalPlayer
- backend/app.go: LocalPlayer field is now player.LocalPlayer interface;
  initMPV/setupMPV renamed to initLocalPlayer/setupLocalPlayer (dispatch
  via build-tagged files)
- backend/app_player_mpv.go: //go:build !localav — extracts initLocalPlayer
  for the mpv backend
- backend/playbackengine.go, ui/**:  update type assertions and imports to
  use player.* types instead of mpv.*
