+++ backend/playbackengine.go
@@ -984,8 +984,8 @@ func (p *playbackEngine) setTrack(idx int, next bool, startTime float64) error {
 					url = filepath
 				}
 			}
-			if mpvP, ok := p.player.(*mpv.Player); ok && !isTrack {
-				mpvP.ObserveIcyRadioTitle(func(icytitle string) {
+			if lpP, ok := p.player.(player.LocalPlayer); ok && !isTrack {
+				lpP.ObserveIcyRadioTitle(func(icytitle string) {
 					var title, artist string
 					if s := strings.Split(icytitle, " - "); len(s) == 2 {
 						title, artist = s[1], s[0]
@@ -997,7 +997,7 @@ func (p *playbackEngine) setTrack(idx int, next bool, startTime float64) error {
 					}
 				})
 			} else if ok {
-				mpvP.UnobserveIcyRadioTitle()
+				lpP.UnobserveIcyRadioTitle()
 			}
 			if url == "" {
 				return errors.New("no stream URL")
