+++ ui/controller/controller.go
@@ -16,7 +16,6 @@ import (
 	"github.com/dweymouth/supersonic/backend"
 	"github.com/dweymouth/supersonic/backend/mediaprovider"
 	"github.com/dweymouth/supersonic/backend/player"
-	"github.com/dweymouth/supersonic/backend/player/mpv"
 	"github.com/dweymouth/supersonic/sharedutil"
 	"github.com/dweymouth/supersonic/ui/dialogs"
 	myTheme "github.com/dweymouth/supersonic/ui/theme"
@@ -312,12 +311,12 @@ func (c *Controller) ShowSettingsDialog(themeUpdateCallbk func(), themeFiles map
 	devs, err := c.App.LocalPlayer.ListAudioDevices()
 	if err != nil {
 		log.Printf("error listing audio devices: %v", err)
-		devs = []mpv.AudioDevice{{Name: "auto", Description: lang.L("Autoselect device")}}
+		devs = []player.AudioDevice{{Name: "auto", Description: lang.L("Autoselect device")}}
 	}
 
 	curPlayer := c.App.PlaybackManager.CurrentPlayer()
 	_, isReplayGainPlayer := curPlayer.(player.ReplayGainPlayer)
-	_, isEqualizerPlayer := curPlayer.(*mpv.Player)
+	_, isEqualizerPlayer := curPlayer.(player.LocalPlayer)
 	_, canSavePlayQueue := c.App.ServerManager.Server.(mediaprovider.CanSavePlayQueue)
 	isLocalPlayer := isEqualizerPlayer
 	bands := c.App.LocalPlayer.Equalizer().BandFrequencies()
@@ -346,9 +345,9 @@ func (c *Controller) ShowSettingsDialog(themeUpdateCallbk func(), themeFiles map
 	dlg.OnThemeSettingChanged = themeUpdateCallbk
 	dlg.OnEqualizerSettingsChanged = func() {
 		// Create the appropriate equalizer type based on config
-		var eq mpv.Equalizer
+		var eq player.Equalizer
 		if c.App.Config.LocalPlayback.EqualizerType == "ISO10Band" {
-			eq10 := &mpv.ISO10BandEqualizer{
+			eq10 := &player.ISO10BandEqualizer{
 				Disabled: !c.App.Config.LocalPlayback.EqualizerEnabled,
 				EQPreamp: c.App.Config.LocalPlayback.EqualizerPreamp,
 			}
@@ -359,7 +358,7 @@ func (c *Controller) ShowSettingsDialog(themeUpdateCallbk func(), themeFiles map
 			}
 			eq = eq10
 		} else {
-			eq15 := &mpv.ISO15BandEqualizer{
+			eq15 := &player.ISO15BandEqualizer{
 				Disabled: !c.App.Config.LocalPlayback.EqualizerEnabled,
 				EQPreamp: c.App.Config.LocalPlayback.EqualizerPreamp,
 			}
